requirements.txt
pillow
termcolor
